# temp_reference/fastapi-backend/app/routers/admin.py
"""
Admin Router
Handles admin operations endpoints
"""

from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.ext.asyncio import AsyncSession
from typing import List, Optional

from app.database import get_db
from app.models.admin import User, UserCreate, UserUpdate, SystemStats, HealthCheck, AdminLog
from app.dependencies import get_current_active_user

router = APIRouter()

@router.get("/users", response_model=List[User])
async def get_users(
    skip: int = 0,
    limit: int = 100,
    search: Optional[str] = None,
    is_active: Optional[bool] = None,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get all users (admin only)"""
    # TODO: Implement user listing logic with admin authorization
    return []

@router.post("/users", response_model=User)
async def create_user(
    user_data: UserCreate,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Create new user (admin only)"""
    # TODO: Implement user creation logic with admin authorization
    return User(
        id=1,
        email=user_data.email,
        username=user_data.username,
        full_name=user_data.full_name,
        is_active=user_data.is_active,
        is_superuser=user_data.is_superuser,
        created_at=datetime.now(),
        updated_at=datetime.now()
    )

@router.get("/users/{user_id}", response_model=User)
async def get_user(
    user_id: int,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get specific user (admin only)"""
    # TODO: Implement user retrieval logic with admin authorization
    raise HTTPException(status_code=404, detail="User not found")

@router.put("/users/{user_id}", response_model=User)
async def update_user(
    user_id: int,
    user_update: UserUpdate,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Update user (admin only)"""
    # TODO: Implement user update logic with admin authorization
    raise HTTPException(status_code=404, detail="User not found")

@router.delete("/users/{user_id}")
async def delete_user(
    user_id: int,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Delete user (admin only)"""
    # TODO: Implement user deletion logic with admin authorization
    return {"message": "User deleted successfully"}

@router.get("/stats", response_model=SystemStats)
async def get_system_stats(
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get system statistics (admin only)"""
    # TODO: Implement system stats logic with admin authorization
    return SystemStats(
        total_users=100,
        active_users=75,
        total_orders=5000,
        total_trades=4500,
        total_volume=1000000.0,
        server_uptime="5 days, 12:30:45",
        memory_usage=65.5,
        cpu_usage=45.2
    )

@router.get("/orders", response_model=List[dict])
async def get_all_orders(
    skip: int = 0,
    limit: int = 100,
    status: Optional[str] = None,
    user_id: Optional[int] = None,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get all orders (admin view)"""
    # TODO: Implement admin order listing logic
    return []

@router.get("/health", response_model=HealthCheck)
async def get_detailed_health(
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get detailed health check (admin only)"""
    # TODO: Implement detailed health check logic
    return HealthCheck(
        status="healthy",
        timestamp=datetime.now(),
        database_status="connected",
        websocket_status="running",
        external_apis=[
            {"name": "Dhan API", "status": "connected", "response_time": "150ms"},
            {"name": "Market Data", "status": "connected", "response_time": "75ms"}
        ]
    )

@router.get("/logs", response_model=List[AdminLog])
async def get_admin_logs(
    skip: int = 0,
    limit: int = 100,
    action: Optional[str] = None,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get admin logs (admin only)"""
    # TODO: Implement admin logs retrieval logic
    return []