# temp_reference/fastapi-backend/app/routers/portfolio.py
"""
Portfolio Router
Handles portfolio operations endpoints
"""

from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.ext.asyncio import AsyncSession
from typing import List, Optional

from app.database import get_db
from app.models.portfolio import Position, Holding, WalletBalance, PortfolioSummary, PnLReport
from app.dependencies import get_current_active_user

router = APIRouter()

@router.get("/positions", response_model=List[Position])
async def get_positions(
    product_type: Optional[str] = None,
    limit: int = 100,
    offset: int = 0,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get user positions"""
    # TODO: Implement positions retrieval logic
    return []

@router.get("/positions/{position_id}")
async def get_position_details(
    position_id: int,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get specific position details"""
    # TODO: Implement position details logic
    raise HTTPException(status_code=404, detail="Position not found")

@router.post("/positions/{position_id}/square-off")
async def square_off_position(
    position_id: int,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Square off position"""
    # TODO: Implement square off logic
    return {"message": "Position squared off successfully"}

@router.get("/holdings", response_model=List[Holding])
async def get_holdings(
    limit: int = 100,
    offset: int = 0,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get user holdings"""
    # TODO: Implement holdings retrieval logic
    return []

@router.get("/balance", response_model=WalletBalance)
async def get_wallet_balance(
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get wallet balance"""
    # TODO: Implement wallet balance logic
    from decimal import Decimal
    return WalletBalance(
        user_id=current_user["id"],
        total_balance=Decimal("100000.00"),
        available_balance=Decimal("80000.00"),
        used_balance=Decimal("20000.00"),
        margin_used=Decimal("5000.00"),
        opening_balance=Decimal("100000.00"),
        realized_pnl=Decimal("1000.00"),
        unrealized_pnl=Decimal("500.00"),
        last_updated=datetime.now()
    )

@router.get("/summary", response_model=PortfolioSummary)
async def get_portfolio_summary(
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get portfolio summary"""
    # TODO: Implement portfolio summary logic
    from decimal import Decimal
    return PortfolioSummary(
        user_id=current_user["id"],
        total_investment=Decimal("50000.00"),
        current_value=Decimal("55000.00"),
        total_pnl=Decimal("5000.00"),
        pnl_percentage=10.0,
        total_positions=5,
        active_positions=3,
        total_holdings=10,
        last_updated=datetime.now()
    )

@router.get("/pnl", response_model=PnLReport)
async def get_pnl_report(
    period: str = "daily",
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get P&L report"""
    # TODO: Implement P&L report logic
    from decimal import Decimal
    return PnLReport(
        user_id=current_user["id"],
        period_start=datetime.now().replace(hour=0, minute=0, second=0, microsecond=0),
        period_end=datetime.now(),
        realized_pnl=Decimal("1000.00"),
        unrealized_pnl=Decimal("500.00"),
        total_pnl=Decimal("1500.00"),
        trades_count=25,
        winning_trades=15,
        losing_trades=10,
        win_rate=60.0
    )