# temp_reference/fastapi-backend/app/routers/trading.py
"""
Trading Router
Handles trading operations endpoints
"""

from fastapi import APIRouter, Depends, HTTPException, status
from sqlalchemy.ext.asyncio import AsyncSession
from typing import List, Optional

from app.database import get_db
from app.models.trading import Order, OrderCreate, OrderUpdate, Trade, BasketOrder
from app.dependencies import get_current_active_user

router = APIRouter()

@router.post("/orders", response_model=Order)
async def place_order(
    order_data: OrderCreate,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Place a new order"""
    # TODO: Implement order placement logic
    return Order(
        id=1,
        user_id=current_user["id"],
        security_id=order_data.security_id,
        quantity=order_data.quantity,
        price=order_data.price,
        order_type=order_data.order_type,
        transaction_type=order_data.transaction_type,
        product_type=order_data.product_type,
        status="PENDING",
        created_at=datetime.now(),
        updated_at=datetime.now()
    )

@router.get("/orders", response_model=List[Order])
async def get_orders(
    status: Optional[str] = None,
    limit: int = 100,
    offset: int = 0,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get user orders"""
    # TODO: Implement order retrieval logic
    return []

@router.get("/orders/{order_id}", response_model=Order)
async def get_order(
    order_id: int,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get specific order details"""
    # TODO: Implement order details retrieval logic
    raise HTTPException(status_code=404, detail="Order not found")

@router.put("/orders/{order_id}", response_model=Order)
async def modify_order(
    order_id: int,
    order_update: OrderUpdate,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Modify existing order"""
    # TODO: Implement order modification logic
    raise HTTPException(status_code=404, detail="Order not found")

@router.delete("/orders/{order_id}")
async def cancel_order(
    order_id: int,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Cancel order"""
    # TODO: Implement order cancellation logic
    return {"message": "Order cancelled successfully"}

@router.get("/trades", response_model=List[Trade])
async def get_trades(
    limit: int = 100,
    offset: int = 0,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Get user trades"""
    # TODO: Implement trade retrieval logic
    return []

@router.post("/basket-orders", response_model=dict)
async def place_basket_order(
    basket_order: BasketOrder,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Place basket order (multiple orders)"""
    # TODO: Implement basket order logic
    return {"message": "Basket order placed successfully", "basket_id": 1}

@router.post("/smart-order", response_model=Order)
async def place_smart_order(
    security_id: str,
    amount: float,
    order_type: str,
    current_user: dict = Depends(get_current_active_user),
    db: AsyncSession = Depends(get_db)
):
    """Place smart order with auto-sizing"""
    # TODO: Implement smart order logic
    return Order(
        id=1,
        user_id=current_user["id"],
        security_id=security_id,
        quantity=100,
        price=100.0,
        order_type="MARKET",
        transaction_type="BUY",
        product_type="INTRADAY",
        status="EXECUTED",
        created_at=datetime.now(),
        updated_at=datetime.now()
    )