# temp_reference/public/Dhan-API_document.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "DhanHQ API"
  },
  "servers": [
    {
      "url": "https://api.dhan.co/v2",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/orders/{order-id}": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "get order by id",
        "description": "The api allows you to retrieve the details of an order placed during the day with their last updated status, with their order id.",
        "externalDocs": {
          "description": "please refer the link for detailed docs",
          "url": "https://dhanhq.co/docs/v1/"
        },
        "operationId": "getorderbyorderid",
        "parameters": [
          {
            "name": "access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Orders"
        ],
        "summary": "modify pending order",
        "description": "The api allows you modify pending order in orderbook. The fields that can be modified are price, quantity, order type & validity.",
        "externalDocs": {
          "description": "please refer the link for detailed docs",
          "url": "https://dhanhq.co/docs/v1/"
        },
        "operationId": "modifyorder",
        "parameters": [
          {
            "name": "access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderModifyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStatusResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Orders"
        ],
        "summary": "cancel given order",
        "description": "The api allows you to cancel a pending order using the order id",
        "externalDocs": {
          "description": "please refer the link for detailed docs",
          "url": "https://dhanhq.co/docs/v1/"
        },
        "operationId": "cancelorder",
        "parameters": [
          {
            "name": "access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "order-id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStatusResponse"
                }
              }
            }
          }
        }
      }
    },
    "/orders": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "get current orders list",
        "description": "The API lets you retrieve an array of all orders requested in a day with their last updated status.",
        "externalDocs": {
          "description": "please refer the link for detailed docs",
          "url": "https://dhanhq.co/docs/v1/"
        },
        "operationId": "getorders",
        "parameters": [
          {
            "name": "access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/OrderResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Orders"
        ],
        "summary": "place an order",
        "description": "The order request API allows you to place new order.",
        "externalDocs": {
          "description": "please refer the link for detailed docs",
          "url": "https://dhanhq.co/docs/v1/"
        },
        "operationId": "placeorder",
        "parameters": [
          {
            "name": "access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderStatusResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/trades": {
      "get": {
        "tags": [
          "Orders"
        ],
        "summary": "get all trades",
        "description": "The api allows you retrieve an array of all trades executed in a day.",
        "externalDocs": {
          "description": "please refer the link for detailed docs",
          "url": "https://dhanhq.co/docs/v1/"
        },
        "operationId": "getalltrades",
        "parameters": [
          {
            "name": "access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TradeResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/positions": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "get current positions",
        "description": "Users can retrieve a list of all open positions for the day. This includes all F&O carryforward positions as well.",
        "externalDocs": {
          "description": "please refer the link for detailed docs",
          "url": "https://dhanhq.co/docs/v1/"
        },
        "operationId": "getpositions",
        "parameters": [
          {
            "name": "access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PositionResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/holdings": {
      "get": {
        "tags": [
          "Portfolio"
        ],
        "summary": "get current holdings",
        "description": "Users can retrieve all holdings bought/sold in previous trading sessions. All T1 and delivered quantities can be fetched.",
        "externalDocs": {
          "description": "please refer the link for detailed docs",
          "url": "https://dhanhq.co/docs/v1/"
        },
        "operationId": "getholdings",
        "parameters": [
          {
            "name": "access-token",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HoldingResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "OrderStatusResponse": {
        "type": "object",
        "properties": {
          "orderId": {
            "type": "string",
            "description": "Order specific identification generated by Dhan"
          },
          "orderStatus": {
            "type": "string",
            "description": "| **Enum Values** | **Description**                             |\n|-----------------|---------------------------------------------|\n| TRANSIT         | Did not reach the exchange server           |\n| PENDING         | Reached at exchange end, awaiting execution |\n| REJECTED        | Rejected at exchange/broker’s end           |\n| CANCELLED       | Cancelled by user                           |\n| PART_TRADED     | Partially Executed                          |\n| TRADED          | Executed                                    |\n| EXPIRED         | Validity of order is expired                |",
            "enum": [
              "TRANSIT",
              "PENDING",
              "REJECTED",
              "CANCELLED",
              "PART_TRADED",
              "TRADED",
              "EXPIRED",
              "MODIFIED",
              "TRIGGERED",
              "INACTIVE"
            ]
          }
        }
      },
      "OrderModifyRequest": {
        "type": "object",
        "properties": {
          "dhanClientId": {
            "type": "string",
            "description": "User specific identification generated by Dhan"
          },
          "orderId": {
            "type": "string",
            "description": "Order specific identification generated by Dhan"
          },
          "orderType": {
            "type": "string",
            "description": "| **Enum Values**  | **Description**             |\n|------------------|-----------------------------|\n| LIMIT            | For Limit Order types       |\n| MARKET           | For market Order types      |\n| STOP_LOSS        | For Stop Loss Limit orders  |\n| STOP_LOSS_MARKET | For Stop Loss Market orders |",
            "enum": [
              "LIMIT",
              "MARKET",
              "STOP_LOSS",
              "STOP_LOSS_MARKET"
            ]
          },
          "legName": {
            "type": "string",
            "description": "Order leg where modification is to be done",
            "enum": [
              "ENTRY_LEG",
              "STOP_LOSS_LEG",
              "TARGET_LEG",
              "NA"
            ]
          },
          "quantity": {
            "type": "integer",
            "description": "Number of shares for the order",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "description": "Price at which order is placed",
            "format": "float"
          },
          "disclosedQuantity": {
            "type": "integer",
            "description": "Number of shares visible (Keep more than 30% of quantity)",
            "format": "int32"
          },
          "triggerPrice": {
            "type": "number",
            "description": "Price at which the order is triggered, in case of SL-M & SL-L",
            "format": "float"
          },
          "validity": {
            "type": "string",
            "description": "| **Enum Values** | **Description**       |\n|-----------------|-----------------------|\n| DAY             | Valid till end of day |\n| IOC             | Immediate or Cancel   |",
            "enum": [
              "DAY",
              "IOC"
            ]
          }
        }
      },
...
      "OrderResponse": {
        "type": "object",
        "properties": {
          "dhanClientId": {
            "type": "string",
            "description": "User specific identification generated by Dhan"
          },
          "orderId": {
            "type": "string",
            "description": "Order specific identification generated by Dhan"
          },
          "exchangeOrderId": {
            "type": "string",
            "description": "Order specific identification generated by Exchange"
          },
          "correlationId": {
            "type": "string",
            "description": "The user/partner generated id for tracking back"
          },
          "orderStatus": {
            "type": "string",
            "description": "| **Enum Values** | **Description**                             |\n|-----------------|---------------------------------------------|\n| TRANSIT         | Did not reach the exchange server           |\n| PENDING         | Reached at exchange end, awaiting execution |\n| REJECTED        | Rejected at exchange/broker’s end           |\n| CANCELLED       | Cancelled by user                           |\n| PART_TRADED     | Partially Executed                          |\n| TRADED          | Executed                                    |\n| EXPIRED         | Validity of order is expired                |",
            "enum": [
              "TRANSIT",
              "PENDING",
              "REJECTED",
              "CANCELLED",
              "PART_TRADED",
              "TRADED",
              "EXPIRED"
            ]
          },
          "transactionType": {
            "type": "string",
            "description": "Signifies the type of transaction whether it's BUY or SELL",
            "enum": [
              "BUY",
              "SELL"
            ]
          },
          "exchangeSegment": {
            "type": "string",
            "description": "| **Enums**    | **Exchange** | **Segment**       |\n|--------------|--------------|-------------------|\n| NSE_EQ       | NSE          | Equity Cash       |\n| NSE_FNO      | NSE          | Futures & Options |\n| BSE_EQ       | BSE          | Equity Cash       |\n| BSE_FNO      | BSE          | Futures & Options |\n| MCX_COMM     | MCX          | Commodity         |\n",
            "enum": [
              "NSE_EQ",
              "NSE_FNO",
              "BSE_EQ",
              "BSE_FNO",
              "MCX_COMM"
            ]
          },
          "productType": {
            "type": "string",
            "description": "| **Enum Values** | **Description**                                |\n|-----------------|------------------------------------------------|\n| CNC             | Cash & Carry for equity deliveries             |\n| INTRADAY        | Intraday for Equity, Futures & Options         |\n| MARGIN          | Carry Forward in Futures & Options             |\n| CO              | Cover Order; entry and stop loss               |\n| BO              | Bracket Order; entry, stop loss & target price |\n| MTF             | Margin Traded Fund                             |\n**CO & BO product types will be valid only for Intraday**",
            "enum": [
              "CNC",
              "INTRADAY",
              "MARGIN",
              "MTF",
              "CO",
              "BO"
            ]
          },
          "orderType": {
            "type": "string",
            "description": "| **Enum Values**  | **Description**             |\n|------------------|-----------------------------|\n| LIMIT            | For Limit Order types       |\n| MARKET           | For market Order types      |\n| STOP_LOSS        | For Stop Loss Limit orders  |\n| STOP_LOSS_MARKET | For Stop Loss Market orders |",
            "enum": [
              "LIMIT",
              "MARKET",
              "STOP_LOSS",
              "STOP_LOSS_MARKET"
            ]
          },
          "validity": {
            "type": "string",
            "description": "| **Enum Values** | **Description**       |\n|-----------------|-----------------------|\n| DAY             | Valid till end of day |\n| IOC             | Immediate or Cancel   |",
            "enum": [
              "DAY",
              "IOC"
            ]
          },
          "tradingSymbol": {
            "type": "string",
            "description": "Exchange standard  trading symbol"
          },
          "securityId": {
            "type": "string",
            "description": "Exchange standard identification for each scrip"
          },
          "quantity": {
            "type": "integer",
            "description": "Number of shares for the order",
            "format": "int32"
          },
          "disclosedQuantity": {
            "type": "integer",
            "description": "Number shares visible in the market depth",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "description": "Price at which the order is requested to execute",
            "format": "float"
          },
          "triggerPrice": {
            "type": "number",
            "description": "Price at which the order is triggered",
            "format": "float"
          },
          "afterMarketOrder": {
            "type": "boolean",
            "description": "Flag to inform that the order placed is After Market Order"
          },
          "boProfitValue": {
            "type": "number",
            "description": "Bracket order Target price",
            "format": "float"
          },
          "boStopLossValue": {
            "type": "number",
            "description": "Bracket Order/Cover Order Stop Loss Price",
            "format": "float"
          },
          "legName": {
            "type": "string",
            "description": "Order leg where modification is to be done",
            "enum": [
              "ENTRY_LEG",
              "STOP_LOSS_LEG",
              "TARGET_LEG",
              "NA"
            ]
          },
          "createTime": {
            "type": "string",
            "description": "Time at which the record is created"
          },
          "updateTime": {
            "type": "string",
            "description": "Time at which the record is updated"
          },
          "exchangeTime": {
            "type": "string",
            "description": "Time at which order reached at exchange end"
          },
          "drvExpiryDate": {
            "type": "string",
            "description": "For F&O, expiry date of contract"
          },
          "drvOptionType": {
            "type": "string",
            "description": "For Options, type CALL or PUT",
            "enum": [
              "CALL",
              "PUT",
              "NA"
            ]
          },
          "drvStrikePrice": {
            "type": "number",
            "description": "For Options, Strike Price",
            "format": "float"
          },
          "omsErrorCode": {
            "type": "string",
            "description": "When order is rejected or failed, omsErrorCode would be populated"
          },
          "omsErrorDescription": {
            "type": "string",
            "description": "When order is rejected or failed, omsErrorDescription would be populated"
          },
          "algoId": {
            "type": "string",
            "description": "Algo_ID"
          },
          "remainingQuantity": {
            "type": "integer",
            "description": "Number of shares yet to be traded for the order",
            "format": "int32"
          },
          "averageTradedPrice": {
            "type": "number",
            "description": "Average price at which order is traded",
            "format": "float"
          },
          "filledQty": {
            "type": "integer",
            "description": "Filled Quantity",
            "format": "int32"
          }
        }
      },
...
      "HoldingResponse": {
        "type": "object",
        "properties": {
          "exchange": {
            "type": "string",
            "description": "Enum values is either NSE or BSE",
            "enum": [
              "NSE",
              "BSE",
              "MCX",
              "ALL"
            ]
          },
          "tradingSymbol": {
            "type": "string",
            "description": "Exchange standard  trading symbol"
          },
          "securityId": {
            "type": "string",
            "description": "Exchange standard identification for each scrip"
          },
          "isin": {
            "type": "string",
            "description": "International Securities Identification Number"
          },
          "totalQty": {
            "type": "integer",
            "description": "Total number of shares in holding for given stock",
            "format": "int32"
          },
          "dpQty": {
            "type": "integer",
            "description": "Quantities present in depository",
            "format": "int32"
          },
          "t1Qty": {
            "type": "integer",
            "description": "Quantities not delivered to depository",
            "format": "int32"
          },
          "mtf_t1_qty": {
            "type": "integer",
            "description": "MTF  Quantities not delivered to depository",
            "format": "int32"
          },
          "mtf_qty": {
            "type": "integer",
            "description": "MTF Quantities delivered to depository",
            "format": "int32"
          },
          "availableQty": {
            "type": "integer",
            "description": "Quantities available for transactions",
            "format": "int32"
          },
          "collateralQty": {
            "type": "integer",
            "description": "Quantities placed as collateral with broker",
            "format": "int32"
          },
          "avgCostPrice": {
            "type": "number",
            "description": "Average Buy Price of total quantities",
            "format": "float"
          },
          "lastTradedPrice": {
            "type": "number",
            "description": "Last Traded Price for the scrip",
            "format": "float"
          }
        }
      },
...
    }
  }
}